Code and Scene Clean Up
No new features, did fix a bug with the appliance is busy script, and also just made that section more dynamic and easy to add to. Lessened the crowding of the scene in the inspector by removing and moving some objects, makin sure to keep the order so things shouldn't break. Updates some other very minor things to but those are the big ones.